requirements.txt
transformers==4.40.0
torch
huggingface_hub
gradio
python-dotenv
smolagents
beautifulsoup4